feat: Add despawning to skeles and balls
