+++ src/game/behaviors/mod.rs
@@ -1,3 +1,4 @@
+pub mod despawn;
 pub mod spawn;
 pub mod target_ent;
 
@@ -11,6 +12,7 @@ pub struct MovementSpeed(pub f32);
 
 #[auto_plugin(app=app)]
 pub(crate) fn plugin(app: &mut App) {
+    app.add_plugins(despawn::plugin);
     app.add_plugins(spawn::plugin);
     app.add_plugins(target_ent::plugin);
 }
